Move ResponseParser and RequestResponseHandler to server package
