+++ src/main/java/httpserver/server/ResponseParser.java
@@ -1,4 +1,4 @@
-package httpserver.http;
+package httpserver.server;
 
 import httpserver.http.response.Response;
 
@@ -7,5 +7,4 @@ public class ResponseParser {
     public static String parse(Response response) {
         return response.toString();
     }
-
 }
