+++ src/test/java/httpserver/http/RequestResponseSenderTest.java
@@ -1,5 +1,6 @@
 package httpserver.http;
 
+import httpserver.server.RequestResponseHandler;
 import org.junit.jupiter.api.Test;
 
 import java.io.BufferedReader;
